# openapi.yml
openapi: 3.1.0
info:
  title: CubeAPI
  description: OpenAPI contract for the CubeSandbox dashboard surface.
  license:
    name: ''
  version: 0.1.0
servers:
- url: /cubeapi/v1
  description: CubeAPI dashboard surface
paths:
  /cluster/overview:
    get:
      tags:
      - handlers::cluster
      operationId: cluster_overview
      responses:
        '200':
          description: Cluster capacity overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterOverview'
        '404':
          description: Cluster endpoint unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /cluster/versions:
    get:
      tags:
      - handlers::cluster
      operationId: cluster_versions
      responses:
        '200':
          description: Cluster component version matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionMatrixView'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /nodes:
    get:
      tags:
      - handlers::cluster
      operationId: list_nodes
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeView'
        '404':
          description: Node inventory unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /nodes/{nodeID}:
    get:
      tags:
      - handlers::cluster
      operationId: get_node
      parameters:
      - name: nodeID
        in: path
        description: Node identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Node detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeView'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - handlers::cluster
      operationId: delete_node
      parameters:
      - name: nodeID
        in: path
        description: Node identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Node removed
        '400':
          description: Invalid node identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 'Removal blocked: node is not isolated/reachable, sandbox creation is in progress, or sandbox specs, active runtime refs, template replicas, or artifact placements remain'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ClusterOverview:
      type: object
      required:
      - nodeCount
      - healthyNodes
      - totalCpuMilli
      - allocatableCpuMilli
      - totalMemoryMB
      - allocatableMemoryMB
      - maxMvmSlots
      properties:
        allocatableCpuMilli:
          type: integer
          format: int64
          description: Currently-allocatable CPU in millicpu.
        allocatableMemoryMB:
          type: integer
          format: int64
          description: Currently-allocatable memory in MiB.
        healthyNodes:
          type: integer
          minimum: 0
        maxMvmSlots:
          type: integer
          format: int64
          description: Sum of every node's maximum MVM slots.
        nodeCount:
          type: integer
          minimum: 0
        totalCpuMilli:
          type: integer
          format: int64
          description: Total CPU across the cluster, expressed in millicpu.
        totalMemoryMB:
          type: integer
          format: int64
          description: Total memory in MiB.
    ComponentMatrixRowView:
      type: object
      description: Per-component aggregation across all nodes.
      required:
      - component
      - consistent
      - versions
      properties:
        component:
          type: string
        consistent:
          type: boolean
        declaredVersion:
          type: string
        declaredVersions:
          type: array
          items:
            type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ComponentVersionGroupView'
    ComponentVersionGroupView:
      type: object
      description: A group of nodes that report the same version of a component.
      required:
      - version
      - nodes
      properties:
        nodes:
          type: array
          items:
            type: string
        version:
          type: string
    ComponentVersionView:
      type: object
      description: One component's version on a node.
      required:
      - component
      properties:
        buildTime:
          type: string
        commit:
          type: string
        component:
          type: string
        source:
          type: string
        version:
          type: string
    ControlPlaneVersionView:
      type: object
      description: Control-plane reference version (the cluster's target version).
      required:
      - version
      properties:
        buildTime:
          type: string
        commit:
          type: string
        version:
          type: string
...
    NodeComponentEntryView:
      type: object
      description: A single component version on a single node, with release declaration membership.
      required:
      - component
      - version
      - declared
      properties:
        component:
          type: string
        declared:
          type: boolean
        version:
          type: string
    NodeConditionView:
      type: object
      required:
      - type
      - status
      properties:
        lastHeartbeatTime:
          type:
          - string
          - 'null'
          format: date-time
        message:
          type: string
        reason:
          type: string
        status:
          type: string
        type:
          type: string
    NodeResourcesView:
      type: object
      required:
      - cpuMilli
      - memoryMB
      properties:
        cpuMilli:
          type: integer
          format: int64
          description: CPU capacity or availability expressed in millicpu.
        memoryMB:
          type: integer
          format: int64
    NodeVersionRowView:
      type: object
      description: Per-node view of the version matrix.
      required:
      - nodeID
      - healthy
      - components
      properties:
        components:
          type: array
          items:
            $ref: '#/components/schemas/NodeComponentEntryView'
        healthy:
          type: boolean
        nodeID:
          type: string
    NodeView:
      type: object
      required:
      - nodeID
      - hostIP
      - healthy
      - capacity
      - allocatable
      - cpuSaturation
      - memorySaturation
      - maxMvmSlots
      - quotaCpu
      - quotaMemMB
      - createConcurrentNum
      properties:
        allocatable:
          $ref: '#/components/schemas/NodeResourcesView'
        capacity:
          $ref: '#/components/schemas/NodeResourcesView'
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/NodeConditionView'
        cpuSaturation:
          type: number
          format: float
          description: Percentage (0-100) of CPU currently in use.
        createConcurrentNum:
          type: integer
          format: int64
          description: Max concurrent sandbox-create operations on this node.
        healthy:
          type: boolean
        heartbeatTime:
          type:
          - string
          - 'null'
          format: date-time
        hostIP:
          type: string
        instanceType:
          type: string
        localTemplates:
          type: array
          items:
            type: string
        maxMvmSlots:
          type: integer
          format: int64
        memorySaturation:
          type: number
          format: float
          description: Percentage (0-100) of memory currently in use.
        nodeID:
          type: string
        quotaCpu:
          type: integer
          format: int64
          description: CPU quota in millicores assigned to this node.
        quotaMemMB:
          type: integer
          format: int64
          description: Memory quota in MiB assigned to this node.
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ComponentVersionView'
...
    VersionMatrixView:
      type: object
      description: Full node x component version matrix.
      required:
      - controlPlane
      - components
      - nodes
      properties:
        components:
          type: array
          items:
            $ref: '#/components/schemas/ComponentMatrixRowView'
        controlPlane:
          $ref: '#/components/schemas/ControlPlaneVersionView'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeVersionRowView'
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...